package: libdrm: declare libdrm-dev as a dependent
The weston depends on libdrm that will require to install libdrm-dev,
and the version of libdrm is not external/libdrm's. The configuration of
weston is not the same as expected.

expect-->
libdrm does not support atomic modesetting, will omit that capability
real-->
libdrm supports atomic API

I don't know whether the external/libdrm supports atomic or not, hack it
for the present.

Change-Id: I4490ad625b1b9eb185310606c6784175f2a239a9
Signed-off-by: Ziyuan Xu <xzy.xu@rock-chips.com>