package: override: provide Qt5-related packages
The 'OpenGL ES' is disabled in official qtbase5-dev, that use 'OpenGL
Desktop' instead(d35641e169dd"Revert the arm64 switch to OpenGL ES.").

Because of the qtbase5-dev has not supported cross compile[1], we have
to create a chroot environment, and use qemu-aarh64-static to cross
compile packages.

1,sudo debootstrap --arch=arm64 buster /media/hdd/debian/ http://deb.debian.org/debian
2,copy qemu-aarch64-static to <target_dir>
3,apt-get source qtbase5-dev
4,apt-get build-dep qtbase5-dev or mk-build-dep qtbase5-dev
5,pick some patches for opengl es and compile workaround.
6,dpkg-buildpackage -us -uc -B -j20

Moreover, the qtwayland and qt5quick5 should be rebuild due to they
rely on qtbase5-dev.

[1]https://www.mail-archive.com/debian-qt-kde@lists.debian.org/msg86810.html

Change-Id: I08deb161119bfa65058084aae2ffe1005be7e84b
Signed-off-by: Ziyuan Xu <xzy.xu@rock-chips.com>